controller/python/requirements.txt
pygame
opencv-python
zeroconf
click
black
traitlets
approxeng.input